create reations between entityContractor, entitySourceFinance

+++ src/main/java/com/cgr/base/infrastructure/repositories/repositories/repositoryEntityProvitionalPlan/IEntityProvitionalPlanJpa.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cgr.base.domain.models.entity.EntityProvitionalPlan;
 
-public interface IEntityProvitionalPlanJpa extends JpaRepository<EntityProvitionalPlan, Long> {
+public interface IEntityProvitionalPlanJpa extends JpaRepository<EntityProvitionalPlan, Integer> {
 
 }
